# Set OptInt32Builder data to nil