The reason we want to do this is so that when pooling the builder as part of the `MultiColSorter` we don't reuse the builder's data for a concurrent existing array.