@asubiotto 2024-02-14
Can you add a regression test? You don't even need concurrency. I would just reuse a builder and check that a previously returned array is intact.

I'm a little concerned by how this will affect scan performance though since we will introduce more allocations. Could you run `BenchmarkQuery` on a prod db?