@asubiotto 2024-02-14
I would also change this for all builders. I think this is a "contract-level" change. i.e. the caller of `NewArray` retains ownership of this array and the builder loses it, so any builder reuse should not affect existing arrays.