@metalmatze 2024-02-16
Here are the results for the `BenchmarkQuery`:
```
name             old time/op  new time/op  delta
Query/Types-24    754µs ± 1%   773µs ± 2%  +2.53%  (p=0.032 n=5+5)
Query/Labels-24   292µs ± 1%   263µs ±10%  -9.79%  (p=0.016 n=4+5)
Query/Values-24   592µs ±20%   628µs ±16%    ~     (p=0.548 n=5+5)
Query/Merge-24    560ms ± 7%   568ms ±10%    ~     (p=1.000 n=5+5)
Query/Range-24   85.3ms ± 5%  87.0ms ±11%    ~     (p=0.310 n=5+5)
```
It doesn't seem too bad, is it?